First implementation of a listener that adds 'NOT' to '@generated' tags
The listener uses the new 'WorkspaceCommandListener' to listen to 'save'
events. Then it parses open Java source files, collects all methods that
are tagged with '@generated' and checks if the body of any of these
methods has been changed since the last save.

For these methods, it opens a dialog that allows the user to decide
whether the 'NOT' tag shall automatically be added to these methods (as
well as an optional explanation text).