Implement the 'org.eclipse.ui.startup' extension point
This is used to activate the AGTele Plugin (and especially to initialize
the listener) when the workbench starts. Although implementing this
extension point is not recommended (see
https://wiki.eclipse.org/FAQ_Can_I_activate_my_plug-in_when_the_workbench_starts%3F)
, this seems to be the only option to ensure that the listener is
already installed when the user performs a save action in a Java editor.
Otherwise, we would have to hope that the AGTele UI Plugin would have
been activated by another way (e.g. by opening one of our editors).