Create separate buttons for 'Push to Ecore' and 'Add NOT Tag'
'Push to Ecore' represents the 'old' behavior (push to ecore -> if this
does not work, add NOT tag) where as 'Add NOT Tag' does not try to push
to ecore first.

Moreover, add another button to the selection dialog that triggers the
creation of a 'TODO' as part of the JavaDoc reminding the user to
incorporate the manual changes back into the Ecore metamodel.

Also change the behavior of the 'automatic' mode: This now adds a NOT
tag (without pushing to ecore) and also creates a TODO.