HandleChangedGeneratedCodeDialog: Remember last entered explanation text
If the user chooses the 'Add NOT tag' option and enters an additional
'explanation text', this text is persisted in the 'DialogSettings'. If
the dialog is launched again, the saved text is reused to pre-populate
the text field.

If a user often reuses the same explanation text, this should save him
lots of time.